Conflicts:
	src/com/jcf/spaceshooter/engine/ShuttleController.java
	src/com/jcf/spaceshooter/screen/GameScreen.java

+++ src/com/jcf/spaceshooter/screen/MainMenuScreen.java
@@ -45,7 +45,7 @@ public class MainMenuScreen extends Screen {
 	}
 
 	@Override
-	public void update(float deltaTime) {
+	public void update(int deltaTime) {
 		Input input = game.getInput();
 		
 		ArrayList<TouchEvent> eventList = input.getTouchEvents();
@@ -68,7 +68,7 @@ public class MainMenuScreen extends Screen {
 	}
 
 	@Override
-	public void present(float deltaTime) {
+	public void present(int deltaTime) {
 		Graphics g = game.getGraphics();
 		
 		g.clear(BGCOLOR);
