+++ src/com/jcf/spaceshooter/screen/Screen.java
@@ -15,8 +15,8 @@ public abstract class Screen {
 		this.game = game;
 	}
 	
-	public abstract void update(float deltaTime);
-	public abstract void present(float deltaTime);
+	public abstract void update(int deltaTime);
+	public abstract void present(int deltaTime);
 	public abstract void pause();
 	public abstract void resume();
 	public abstract void dispose();
